+++ src/ops/array_ops.ts
@@ -25,10 +25,10 @@ import * as tensor_util from '../tensor_util';
 import {ArrayData, DataType, DataTypeMap, Rank, ShapeMap, TensorLike, TensorLike1D, TensorLike2D, TensorLike3D, TensorLike4D, TypedArray} from '../types';
 import * as util from '../util';
 import {parseAxisParam} from './axis_util';
-import {CompareOps} from './compare';
 import {ConcatOps} from './concat';
 import {operation} from './operation';
 import {MPRandGauss} from './rand';
+import {ReductionOps} from './reduction_ops';
 
 export class ArrayOps {
   /**
@@ -892,7 +892,8 @@ export class ArrayOps {
     axis = parseAxisParam(axis, x.shape)[0];
     const grad = (dy: T) => {
       const derX = () => {
-        return ArrayOps.unsortedSegmentSum(dy, indices, x.shape[axis], axis);
+        return ReductionOps.unsortedSegmentSum(
+            dy, indices, x.shape[axis], axis);
       };
       return {x: derX};
     };
@@ -1248,61 +1249,6 @@ export class ArrayOps {
   @doc({heading: 'Tensors', subheading: 'Creation'})
   static print<T extends Tensor>(x: T, verbose = false): void {
     console.log(tensor_util.tensorToString(x, verbose));
-  }
-
-  /**
-   * Computes the sum along segments of a `Tensor`.
-   *
-   * ```js
-   * const x = tf.tensor1d([1, 2, 3, 4]);
-   * const indices = tf.tensor1d([1, 2, 0, 1]);
-   * comst numSegments = 3;
-   * const axis = 0;
-   *
-   * x.unsortedSegmentSum(indices, numSegments, axis).print() //or
-   * tf.unsortedSegmentSum(x, indices, numSegments, axis)
-   * ```
-   * @param x The `Tensor` that will be summed along its segments
-   * @param segmentIds A `Tensor1D` whose rank is equal to the rank of `x`'s
-   * dimension akong the `axis`.  Maps each element of `x` to a segment.
-   * @param numSegments The number of distinct `segmentIds`
-   * @param axis The dimension along which the sums will be
-   * calculated. Defaults to 0.
-   */
-  @doc({heading: 'Operations', subheading: 'Reduction'})
-  static unsortedSegmentSum<T extends Tensor>(
-      x: T, segmentIds: Tensor1D, numSegments: number, axis = 0): T {
-    util.assertArgumentsAreTensors({x, segmentIds}, 'unsortedSegmentSum');
-
-    util.assert(
-        segmentIds.dtype === 'int32', 'Segment Ids must be of dtype `int32`');
-
-    axis = parseAxisParam(axis, x.shape)[0];
-    const res = [];
-    const [dim] = segmentIds.shape;
-
-    // Reshape the segment id's so that they can be broadcast with
-    // x. The new shape should be [1, 1, ... 1, dim, 1, ..., 1] where
-    // dim is at index = axis.
-    const newShape = [];
-    for (let i = 0; i < x.shape.length; i++) {
-      if (i === axis) {
-        newShape.push(dim);
-      } else {
-        newShape.push(1);
-      }
-    }
-
-    const reshapedSegmentIds = ArrayOps.reshape(segmentIds, newShape);
-    for (let i = 0; i < numSegments; i++) {
-      const segmentId = ArrayOps.scalar(i, 'int32');
-      const mask =
-          CompareOps.equal(segmentId, reshapedSegmentIds).asType('float32');
-      const sum = mask.mul(x).sum(axis);
-      res.push(sum);
-    }
-
-    return ArrayOps.stack(res, axis) as T;
   }
 }
