+++ src/ops/array_ops_test.ts
@@ -2286,69 +2286,3 @@ describeWithFlags('expandDims', ALL_ENVS, () => {
         .toThrowError(/Argument 'x' passed to 'expandDims' must be a Tensor/);
   });
 });
-
-describeWithFlags('unsortedSegmentSum', ALL_ENVS, () => {
-  it('tensor1D', () => {
-    const t = tf.tensor1d([1, 2, 3, 4]);
-    const segmentIds = tf.tensor1d([0, 2, 0, 1], 'int32');
-    const numSegments = 3;
-    const res = tf.unsortedSegmentSum(t, segmentIds, numSegments);
-
-    expect(res.shape).toEqual([3]);
-    expectArraysClose(res, [4, 4, 2]);
-  });
-
-  it('tensor2D axis=0', () => {
-    const t = tf.tensor2d([1, 2, 3, 4], [2, 2]);
-    const segmentIds = tf.tensor1d([0, 0], 'int32');
-    const numSegments = 2;
-    const res = tf.unsortedSegmentSum(t, segmentIds, numSegments);
-
-    expect(res.shape).toEqual([2, 2]);
-    expectArraysClose(res, [4, 6, 0, 0]);
-  });
-
-  it('tensor2D axis=11', () => {
-    const t = tf.tensor2d([1, 2, 3, 4], [2, 2]);
-    const segmentIds = tf.tensor1d([0, 0], 'int32');
-    const numSegments = 2;
-    const axis = 1;
-    const res = tf.unsortedSegmentSum(t, segmentIds, numSegments, axis);
-
-    expect(res.shape).toEqual([2, 2]);
-    expectArraysClose(res, [3, 0, 7, 0]);
-  });
-
-  it('tensor3D axis=0', () => {
-    const t = tf.tensor3d([1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12], [3, 2, 2]);
-    const segmentIds = tf.tensor1d([2, 1, 2], 'int32');
-    const numSegments = 3;
-    const axis = 0;
-    const res = tf.unsortedSegmentSum(t, segmentIds, numSegments, axis);
-
-    expect(res.shape).toEqual([3, 2, 2]);
-    expectArraysClose(res, [0, 0, 0, 0, 5, 6, 7, 8, 10, 12, 14, 16]);
-  });
-
-  it('tensor3D axis=1', () => {
-    const t = tf.tensor3d([1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12], [2, 3, 2]);
-    const segmentIds = tf.tensor1d([2, 2, 2], 'int32');
-    const numSegments = 3;
-    const axis = 1;
-    const res = tf.unsortedSegmentSum(t, segmentIds, numSegments, axis);
-
-    expect(res.shape).toEqual([2, 3, 2]);
-    expectArraysClose(res, [0, 0, 0, 0, 9, 12, 0, 0, 0, 0, 27, 30]);
-  });
-
-  it('tensor3D axis=2', () => {
-    const t = tf.tensor3d([1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12], [2, 3, 2]);
-    const segmentIds = tf.tensor1d([0, 0], 'int32');
-    const numSegments = 2;
-    const axis = 2;
-    const res = tf.unsortedSegmentSum(t, segmentIds, numSegments, axis);
-
-    expect(res.shape).toEqual([2, 3, 2]);
-    expectArraysClose(res, [3, 0, 7, 0, 11, 0, 15, 0, 19, 0, 23, 0]);
-  });
-});
